+++ Sources/SuperwallKit/Superwall.swift
@@ -440,10 +440,10 @@ public final class Superwall: NSObject, ObservableObject {
 
   /// Gets an array of all confirmed experiment assignments.
   ///
-  /// - Returns: An array of ``ConfirmedAssignment`` objects.
-  public func getAssignments() -> Set<Assignment> {
-    let confirmedAssignments = dependencyContainer.storage.getAssignments()
-    return confirmedAssignments
+  /// - Returns: An array of ``Assignment`` objects.
+  public func getAssignments() -> [Assignment] {
+    let assignments = dependencyContainer.storage.getAssignments()
+    return Array(assignments)
   }
 
   /// Confirms all experiment assignments and returns them in an array.
@@ -453,8 +453,8 @@ public final class Superwall: NSObject, ObservableObject {
   /// Note that the assignments may be different when a placement is registered due to changes
   /// in user, placement, or device parameters used in audience filters.
   ///
-  /// - Returns: An array of ``ConfirmedAssignment`` objects.
-  public func confirmAllAssignments() async ->Set<Assignment> {
+  /// - Returns: An array of ``Assignment`` objects.
+  public func confirmAllAssignments() async -> [Assignment] {
     let confirmAllAssignments = InternalSuperwallPlacement.ConfirmAllAssignments()
     await track(confirmAllAssignments)
 
@@ -482,7 +482,7 @@ public final class Superwall: NSObject, ObservableObject {
         assignments.update(with: assignment)
       }
     }
-    return assignments
+    return Array(assignments)
   }
 
   /// Confirms all experiment assignments and returns them in an array.
@@ -492,8 +492,8 @@ public final class Superwall: NSObject, ObservableObject {
   /// Note that the assignments may be different when a placement is registered due to changes
   /// in user, placement, or device parameters used in audience filters.
   ///
-  /// - Parameter completion: A completion block that accepts an array of ``ConfirmedAssignment`` objects.
-  public func confirmAllAssignments(completion: ((Set<Assignment>) -> Void)? = nil) {
+  /// - Parameter completion: A completion block that accepts an array of ``Assignment`` objects.
+  public func confirmAllAssignments(completion: (([Assignment]) -> Void)? = nil) {
     Task {
       let result = await confirmAllAssignments()
       completion?(result)