+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 The changelog for `SuperwallKit`. Also see the [releases](https://github.com/superwall/Superwall-iOS/releases) on GitHub.
 
+## 4.0.1
+
+### Enhancements
+
+- Confirms all paywall assignments locally to reduce the amount of preloading of paywalls on each cold app open.
+- Migrates documents used for user and app data out of the documents folder and into the application support folder.
+
 ## 4.0.0
 
 ### Breaking Changes